requirements.txt
Celery
gevent
celery[redis]
celery[gevent]
python-dotenv
polling
